Move functions from process to file
Move filesystem-related functions from process.[ch] to file.[ch].
